move telemetry-permission-dialog to src/common/components
