+++ src/common/components/telemetry-permission-dialog.tsx
@@ -11,7 +11,7 @@ import { PrimaryButton } from 'office-ui-fabric-react/lib/Button';
 import { Checkbox } from 'office-ui-fabric-react/lib/Checkbox';
 import { Dialog, DialogFooter, DialogType } from 'office-ui-fabric-react/lib/Dialog';
 import * as React from 'react';
-import { UserConfigMessageCreator } from '../../common/message-creators/user-config-message-creator';
+import { UserConfigMessageCreator } from '../message-creators/user-config-message-creator';
 
 export interface TelemetryPermissionDialogState {
     isEnableTelemetryChecked: boolean;
